[IMP] point_of_sale: add "Cancel Order" action
In the point_of_sale a user can cancel an order from the ticket_screen.
The plan is to to replace the ticket_screen with the standard list view.
This means that orders that are not yet synced with the server will not
appear there, so there will be no way for the user to cancel them.

In this commit we create a new action button, called "Cancel order".

Task: 4104387
