+++ addons/point_of_sale/static/src/app/screens/product_screen/control_buttons/control_buttons.xml
@@ -39,6 +39,9 @@
                     <t t-if="currentOrder?.fiscal_position_id" t-esc="currentOrder.fiscal_position_id.display_name" />
                     <t t-else="">Tax</t>
                 </button>
+                <button class="btn btn-secondary btn-lg py-5" t-on-click="() => this.pos.onDeleteOrder(this.pos.get_order())">
+                    <i class="fa fa-trash me-1" role="img" /> Cancel Order 
+                </button>
             </div>
         </t>
     </t>
