+++ addons/pos_restaurant/static/src/overrides/components/ticket_screen/ticket_screen.js
@@ -48,19 +48,6 @@ patch(TicketScreen.prototype, {
         await this.pos.setTable(orderTable, order.uuid);
         this.closeTicketScreen();
     },
-    async onDeleteOrder(order) {
-        const confirmed = await super.onDeleteOrder(...arguments);
-        if (
-            confirmed &&
-            this.pos.config.module_pos_restaurant &&
-            this.pos.selectedTable &&
-            !this.pos.models["pos.order"].some(
-                (order) => order.table_id?.id === this.pos.selectedTable.id
-            )
-        ) {
-            return this.pos.showScreen("FloorScreen");
-        }
-    },
     get allowNewOrders() {
         return this.pos.config.module_pos_restaurant
             ? Boolean(this.pos.selectedTable)
